requirements.txt
accelerate==0.33.0
altair==5.3.0
attrs==24.1.0
bitsandbytes==0.43.3
black==24.8.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
filelock==3.13.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.24.5
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.4
nvidia-cuda-runtime-cu12==12.6.37
nvidia-pyindex==1.0.9
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
plotly==5.23.0
protobuf==5.27.3
psutil==6.0.0
pyarrow==17.0.0
pydeck==0.9.1
pydocstring==0.2.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
rpds-py==0.19.1
ruff==0.5.7
safetensors==0.4.4
setuptools==72.1.0
six==1.16.0
smmap==5.0.1
streamlit==1.37.1
sympy==1.12
tbb==2021.11.0
tenacity==8.5.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.0+cu121
torchaudio==2.3.0+cu121
torchvision==0.18.0+cu121
tornado==6.4.1
tqdm==4.66.5
transformers==4.44.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
wheel==0.44.0
